6756771: com.sun.net.httpserver.HttpServer should handle POSTs larger than 2Gig
Update implementation to use long instead of int

+++ jdk/src/share/classes/sun/net/httpserver/ServerImpl.java
@@ -470,13 +470,13 @@ class ServerImpl implements TimeSource {
                 String version = requestLine.substring (start);
                 Headers headers = req.headers();
                 String s = headers.getFirst ("Transfer-encoding");
-                int clen = 0;
+                long clen = 0L;
                 if (s !=null && s.equalsIgnoreCase ("chunked")) {
-                    clen = -1;
+                    clen = -1L;
                 } else {
                     s = headers.getFirst ("Content-Length");
                     if (s != null) {
-                        clen = Integer.parseInt (s);
+                        clen = Long.parseLong(s);
                     }
                 }
                 ctx = contexts.findContext (protocol, uri.getPath());
